chatbot/pyproject.toml
[project]
name = "hanzi-chatbot"
version = "0.1.0"
description = "Chatbot codebase"
requires-python = ">=3.12,<3.13"

dependencies = [
    "altair",
    "GitPython",
    "google-ai-generativelanguage",
    "groq",
    "ipython",
    "langchain",
    "langchain-google-genai",
    "langchain-groq",
    "langfuse",
    "langgraph",
    "langsmith",
    "numpy",
    "fastmcp",
    "openai",
    "pandas",
    "pillow",
    "pydantic",
    "python-dotenv",
    "PyYAML",
    "requests",
    "streamlit",
    "tqdm",
    "watchdog",
]